+++ prompt.c
@@ -263,3 +263,16 @@ prompt_erase_after_cursor(struct prompt *prompt)
     prompt->text[prompt->cursor] = U'\0';
     return true;
 }
+
+bool
+prompt_erase_before_cursor(struct prompt *prompt)
+{
+    if (prompt->cursor == 0) {
+      return false;
+    }
+    memmove(&prompt->text[0],
+            &prompt->text[prompt->cursor],
+            (c32len(prompt->text) - prompt->cursor + 1) * sizeof(char32_t));
+    prompt->cursor = 0;
+    return true;
+}