+++ prompt.h
@@ -27,3 +27,4 @@ bool prompt_erase_prev_char(struct prompt *prompt);
 bool prompt_erase_next_word(struct prompt *prompt);
 bool prompt_erase_prev_word(struct prompt *prompt);
 bool prompt_erase_after_cursor(struct prompt *prompt);
+bool prompt_erase_before_cursor(struct prompt *prompt);