+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 * `include=<path>` option to `fuzzel.ini` ([#205][205]).
 * New option `--list-executables-in-path` add executables presents in
   the $PATH variable to the list ([#284][284]).
+* New key binding: `delete-line-backward` which corresponds to C-u in bash.
+  Also renames `delete-line` to `delete-line-forward` ([#307][307]).
 
 [130]: https://codeberg.org/dnkl/fuzzel/issues/130
 [264]: https://codeberg.org/dnkl/fuzzel/issues/264
@@ -47,6 +49,7 @@
 [237]: https://codeberg.org/dnkl/fuzzel/issues/237
 [205]: https://codeberg.org/dnkl/fuzzel/issues/205
 [284]: https://codeberg.org/dnkl/fuzzel/pulls/284
+[307]: https://codeberg.org/dnkl/fuzzel/pulls/307
 
 
 ### Changed
@@ -56,10 +59,12 @@
 * Exact matches (of the application title) are now sorted first
   ([#259][259]).
 * Set default log level to warning ([#266][266]).
+* Rename `delete-line` binding to `delete-line-forward` ([#307][307]).
 
 [255]: https://codeberg.org/dnkl/fuzzel/issues/255
 [259]: https://codeberg.org/dnkl/fuzzel/issues/259
 [266]: https://codeberg.org/dnkl/fuzzel/pulls/266
+[307]: https://codeberg.org/dnkl/fuzzel/pulls/307
 
 
 ### Deprecated